Disable NodeManager vmem check
By default, we allocate 1GB of physical memory for map and reduce
tasks (mapred.x.memory.mb in mapred-site.xml). We also have a
vmem:pmem ratio of 2:1, meaning our virtual memory usage cannot
exceed 2GB. This is not enough for java8, and vmem frequently goes
above 2GB, causing the compute-slave container to be killed. Disabling
vmem-check-enabled fixes this, and has precedent. See google and #6
here:

http://blog.cloudera.com/blog/2014/04/apache-hadoop-yarn-avoiding-6-time-consuming-gotchas/

A better fix would be to adjust memory.mb on our compute slaves to
match what the system is capable of, but that will be easier to do
when we move our charms to layers.