move dfs config to plugin charm
The dfs replication and blocksize config opts were not being
honored on the namenode charm. This is because clients will
use the hadoop defaults or local hdfs-site.xml configs. Move these
to the plugin, since that's the most likely endpoint for people
to create new files (and hence, utilize these config opts).